Petrel: Remove dead legacy AuthenticationService actor
The 4163-line `AuthenticationService` actor was internal, never
constructed anywhere in Sources/, and unreferenced by Catbird. The
strategy stack (AuthManager + PublicOAuthStrategy / CABOAuthStrategy /
ConfidentialGatewayStrategy / LegacyPasswordStrategy, all composing
OAuthCore) has fully replaced it.

Extracted into the new Sources/Petrel/Auth/AuthTypes.swift, byte-identical,
the live types the file also happened to host:
  - AuthProgressEvent, AuthProgressDelegate, AuthFailureDelegate
  - TokenRefreshResult, ClientAssertionBackendError, AuthError
  - TokenResponse (decoded by TokenRefreshCoordinator and both OAuth strategies)
  - Data.base64URLEscaped() (used by OAuthCore)

Also removed:
  - AuthServiceProtocol — the actor was its only conformer and only reference.
  - The file-private DPoPPayload / PARResponse copies, which OAuthCore
    already redeclares for itself.
  - The invalid-audience branch in NetworkService.performRequest: it existed
    solely to feed `authProvider as? AuthenticationService`, a downcast that
    can never succeed once the actor is gone (and in practice already never
    matched, since the provider is always AuthManager).
  - OAuthCore.nextRefreshResourceOverride, the strategy-side remnant of that
    same one-shot override. Nothing ever read or wrote it. Reimplementing
    invalid_audience recovery on the strategy path is a deliberate follow-up,
    not something this field was ever close to providing.

TokenRefreshCoordinator and RefreshCircuitBreaker are kept — OAuthCore and
the strategies still use both.
