+++ Sources/Petrel/Network/NetworkService.swift
@@ -1402,29 +1402,6 @@ public actor NetworkService: NetworkServiceProtocol {
                         LogManager.logInfo(
                             "Network Service - Attempting standard token refresh/handling for 401 on \(url.absoluteString)."
                         )
-                        // If the 401 is due to invalid audience, set a one-shot resource override
-                        do {
-                            struct OAuthErrorResponse: Decodable { let error: String; let errorDescription: String?; enum CodingKeys: String, CodingKey { case error; case errorDescription = "error_description" } }
-                            var isInvalidAudience = false
-                            // First try JSON body (authorization server style)
-                            if let err = try? jsonDecoder.decode(OAuthErrorResponse.self, from: decompressedData) {
-                                let desc = err.errorDescription?.lowercased() ?? ""
-                                isInvalidAudience = (err.error == "invalid_audience") || (err.error == "invalid_token" && desc.contains("invalid audience"))
-                            }
-                            // Also try WWW-Authenticate header from resource server
-                            if !isInvalidAudience, let www = httpResponse.value(forHTTPHeaderField: "WWW-Authenticate")?.lowercased() {
-                                // Example: DPoP error="invalid_token", error_description="invalid audience"
-                                if www.contains("error=\"invalid_token\""), www.contains("invalid audience") { isInvalidAudience = true }
-                                if www.contains("error=\"invalid_audience\"") { isInvalidAudience = true }
-                            }
-                            if isInvalidAudience, let scheme = url.scheme, let host = url.host {
-                                let resource = "\(scheme)://\(host)"
-                                LogManager.logInfo("Network Service - Preparing one-shot refresh for resource: \(resource)")
-                                if let svc = self.authProvider as? AuthenticationService {
-                                    await svc.setNextRefreshResourceOverride(resource)
-                                }
-                            }
-                        }
                         do {
                             // Use the authProvider's handler (which should attempt refresh)
                             let (retryData, retryResponse) = try await authProvider.handleUnauthorizedResponse(
