+++ Tests/PetrelTests/Auth/RefreshTokenReliabilityTests.swift
@@ -149,259 +149,6 @@ final class MockDIDResolver: DIDResolving, @unchecked Sendable {
     }
 }
 
-private final class AsyncLatch: @unchecked Sendable {
-    private let lock = NSLock()
-    private var signaled = false
-    private var waiters: [CheckedContinuation<Void, Never>] = []
-
-    func signal() {
-        lock.lock()
-        guard !signaled else {
-            lock.unlock()
-            return
-        }
-        signaled = true
-        let currentWaiters = waiters
-        waiters.removeAll()
-        lock.unlock()
-        currentWaiters.forEach { $0.resume() }
-    }
-
-    func wait() async {
-        await withCheckedContinuation { continuation in
-            lock.lock()
-            if signaled {
-                lock.unlock()
-                continuation.resume()
-            } else {
-                waiters.append(continuation)
-                lock.unlock()
-            }
-        }
-    }
-}
-
-private final class RefreshStorageObserver: @unchecked Sendable {
-    let refreshCleared = AsyncLatch()
-    let postClearSessionLookup = AsyncLatch()
-
-    private let lock = NSLock()
-    private let did: String
-    private var didClearRefresh = false
-
-    init(did: String) {
-        self.did = did
-    }
-
-    func observe(_ operation: InMemorySecureStorage.Operation, key: String) {
-        let signals = lock.withLock { () -> (refreshCleared: Bool, sessionLookup: Bool) in
-            switch operation {
-            case .delete where key == "refresh.inProgress.\(did)":
-                didClearRefresh = true
-                return (true, false)
-            case .retrieve where didClearRefresh && key == "session.pending.\(did)":
-                return (false, true)
-            default:
-                return (false, false)
-            }
-        }
-        if signals.refreshCleared {
-            refreshCleared.signal()
-        }
-        if signals.sessionLookup {
-            postClearSessionLookup.signal()
-        }
-    }
-}
-
-private actor AuthenticationRefreshTransport {
-    enum Response {
-        case invalidGrant
-        case serverFailure
-        case unauthorized
-        case refreshSuccess
-        case conflict
-    }
-
-    private let responses: [Response]
-    private var requestCount = 0
-    private var requestWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
-    private var releasePermits = 0
-    private var releaseWaiters: [CheckedContinuation<Void, Never>] = []
-
-    init(responses: [Response]) {
-        self.responses = responses
-    }
-
-    func waitForRequestCount(_ count: Int) async {
-        guard requestCount < count else { return }
-        await withCheckedContinuation { continuation in
-            requestWaiters.append((count, continuation))
-        }
-    }
-
-    func releaseNextRequest() {
-        if let continuation = releaseWaiters.first {
-            releaseWaiters.removeFirst()
-            continuation.resume()
-        } else {
-            releasePermits += 1
-        }
-    }
-
-    func observedRequestCount() -> Int {
-        requestCount
-    }
-
-    func handle(_ box: AuthenticationRefreshURLProtocolBox) async {
-        let responseIndex = requestCount
-        requestCount += 1
-        let readyWaiters = requestWaiters.filter { requestCount >= $0.count }
-        requestWaiters.removeAll { requestCount >= $0.count }
-        readyWaiters.forEach { $0.continuation.resume() }
-
-        if releasePermits > 0 {
-            releasePermits -= 1
-        } else {
-            await withCheckedContinuation { releaseWaiters.append($0) }
-        }
-
-        guard !Task.isCancelled else { return }
-        guard responseIndex < responses.count else {
-            box.protocolInstance.client?.urlProtocol(
-                box.protocolInstance,
-                didFailWithError: URLError(.badServerResponse)
-            )
-            return
-        }
-
-        switch responses[responseIndex] {
-        case .invalidGrant:
-            let response = HTTPURLResponse(
-                url: box.request.url!,
-                statusCode: 400,
-                httpVersion: "HTTP/1.1",
-                headerFields: ["Content-Type": "application/json"]
-            )!
-            let data = Data(#"{"error":"invalid_grant","error_description":"refresh token rejected"}"#.utf8)
-            box.protocolInstance.client?.urlProtocol(
-                box.protocolInstance,
-                didReceive: response,
-                cacheStoragePolicy: .notAllowed
-            )
-            box.protocolInstance.client?.urlProtocol(box.protocolInstance, didLoad: data)
-            box.protocolInstance.client?.urlProtocolDidFinishLoading(box.protocolInstance)
-        case .serverFailure:
-            let response = HTTPURLResponse(
-                url: box.request.url!,
-                statusCode: 400,
-                httpVersion: "HTTP/1.1",
-                headerFields: ["Content-Type": "application/json"]
-            )!
-            let data = Data(#"{"error":"temporarily_unavailable","error_description":"refresh unavailable"}"#.utf8)
-            box.protocolInstance.client?.urlProtocol(
-                box.protocolInstance,
-                didReceive: response,
-                cacheStoragePolicy: .notAllowed
-            )
-            box.protocolInstance.client?.urlProtocol(box.protocolInstance, didLoad: data)
-            box.protocolInstance.client?.urlProtocolDidFinishLoading(box.protocolInstance)
-        case .unauthorized:
-            respond(
-                to: box,
-                statusCode: 401,
-                data: Data(#"{"error":"ExpiredToken"}"#.utf8)
-            )
-        case .refreshSuccess:
-            respond(
-                to: box,
-                statusCode: 200,
-                data: Data(
-                    #"{"accessJwt":"access-refreshed","refreshJwt":"refresh-refreshed","handle":"test.example","did":"did:plc:refreshreliability"}"#.utf8
-                )
-            )
-        case .conflict:
-            respond(
-                to: box,
-                statusCode: 409,
-                data: Data(#"{"error":"DestinationExists"}"#.utf8)
-            )
-        }
-    }
-
-    private func respond(
-        to box: AuthenticationRefreshURLProtocolBox,
-        statusCode: Int,
-        data: Data
-    ) {
-        let response = HTTPURLResponse(
-            url: box.request.url!,
-            statusCode: statusCode,
-            httpVersion: "HTTP/1.1",
-            headerFields: ["Content-Type": "application/json"]
-        )!
-        box.protocolInstance.client?.urlProtocol(
-            box.protocolInstance,
-            didReceive: response,
-            cacheStoragePolicy: .notAllowed
-        )
-        box.protocolInstance.client?.urlProtocol(box.protocolInstance, didLoad: data)
-        box.protocolInstance.client?.urlProtocolDidFinishLoading(box.protocolInstance)
-    }
-}
-
-private final class AuthenticationRefreshURLProtocol: URLProtocol {
-    struct Installation {
-        let host: String
-        let transport: AuthenticationRefreshTransport
-    }
-
-    private static let installation = Mutex<Installation?>(nil)
-    private var loadingTask: Task<Void, Never>?
-
-    static func install(host: String, transport: AuthenticationRefreshTransport) {
-        installation.withLock { $0 = Installation(host: host, transport: transport) }
-    }
-
-    static func uninstall() {
-        installation.withLock { $0 = nil }
-    }
-
-    override class func canInit(with request: URLRequest) -> Bool {
-        installation.withLock { $0?.host == request.url?.host }
-    }
-
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-        request
-    }
-
-    override func startLoading() {
-        guard let transport = Self.installation.withLock({ $0?.transport }) else {
-            client?.urlProtocol(self, didFailWithError: URLError(.unsupportedURL))
-            return
-        }
-        let box = AuthenticationRefreshURLProtocolBox(protocolInstance: self, request: request)
-        loadingTask = Task {
-            await transport.handle(box)
-        }
-    }
-
-    override func stopLoading() {
-        loadingTask?.cancel()
-        loadingTask = nil
-    }
-}
-
-private final class AuthenticationRefreshURLProtocolBox: @unchecked Sendable {
-    let protocolInstance: AuthenticationRefreshURLProtocol
-    let request: URLRequest
-
-    init(protocolInstance: AuthenticationRefreshURLProtocol, request: URLRequest) {
-        self.protocolInstance = protocolInstance
-        self.request = request
-    }
-}
-
 // MARK: - Fixtures
 
 private let testDID = "did:plc:refreshreliability"
@@ -439,45 +186,6 @@ private func makeCore(storage: KeychainStorage) async -> OAuthCore {
         oauthConfig: config,
         didResolver: MockDIDResolver()
     )
-}
-
-private func makeAuthenticationService(
-    storage: KeychainStorage,
-    transport: AuthenticationRefreshTransport,
-    host: String
-) -> AuthenticationService {
-    makeAuthenticationServiceFixture(
-        storage: storage,
-        transport: transport,
-        host: host
-    ).service
-}
-
-private func makeAuthenticationServiceFixture(
-    storage: KeychainStorage,
-    transport: AuthenticationRefreshTransport,
-    host: String
-) -> (service: AuthenticationService, networkService: NetworkService) {
-    let account = Account(
-        did: testDID,
-        handle: "test.example",
-        pdsURL: URL(string: "https://\(host)")!
-    )
-    NetworkService.setNetworkTestProtocolClasses([AuthenticationRefreshURLProtocol.self])
-    let networkService = NetworkService(baseURL: account.pdsURL)
-    AuthenticationRefreshURLProtocol.install(host: host, transport: transport)
-    let service = AuthenticationService(
-        storage: storage,
-        accountManager: MockAccountManager(account: account),
-        networkService: networkService,
-        oauthConfig: OAuthConfig(
-            clientId: "test-client",
-            redirectUri: "test://callback",
-            scope: "atproto"
-        ),
-        didResolver: MockDIDResolver()
-    )
-    return (service, networkService)
 }
 
 /// Runs `body` with an injected in-memory storage backend, always restoring the
@@ -559,141 +267,6 @@ struct RefreshTokenReliabilityTests {
             )
 
             #expect(try await storage.getDPoPKeyRepresentation(for: did) == replacement)
-        }
-    }
-
-    @Test("Named HTTP error path survives AuthenticationService 401 refresh retry")
-    func namedHTTPErrorPathSurvivesAuthenticationServiceRetry() async throws {
-        let backend = InMemorySecureStorage()
-        try await withInMemoryBackend(backend) {
-            let host = "typed-error-auth-\(UUID().uuidString.lowercased()).example"
-            let storage = KeychainStorage(
-                namespace: "test.refresh.authentication.typed-error"
-            )
-            let session = makeSession(
-                refreshToken: "refresh-original",
-                tokenType: .bearer
-            )
-            let transport = AuthenticationRefreshTransport(
-                responses: [.unauthorized, .refreshSuccess, .conflict]
-            )
-            defer {
-                AuthenticationRefreshURLProtocol.uninstall()
-                NetworkService.setNetworkTestProtocolClasses(nil)
-            }
-
-            try await storage.saveAccountAndSession(
-                makeAccount(),
-                session: session,
-                for: testDID
-            )
-            let fixture = makeAuthenticationServiceFixture(
-                storage: storage,
-                transport: transport,
-                host: host
-            )
-            await fixture.networkService.setAuthenticationProvider(fixture.service)
-            let request = URLRequest(
-                url: URL(string: "https://\(host)/xrpc/blue.moji.test")!
-            )
-            let operation = Task {
-                try await fixture.networkService
-                    .performRequestReturningHTTPErrorResponses(
-                        request,
-                        skipTokenRefresh: false
-                    )
-            }
-
-            for requestCount in 1 ... 3 {
-                await transport.waitForRequestCount(requestCount)
-                await transport.releaseNextRequest()
-            }
-
-            let (data, response) = try await operation.value
-            let storedSession = try await storage.getSession(for: testDID)
-            let requestCount = await transport.observedRequestCount()
-            #expect(response.statusCode == 409)
-            #expect(
-                String(decoding: data, as: UTF8.self) ==
-                    #"{"error":"DestinationExists"}"#
-            )
-            #expect(storedSession?.accessToken == "access-refreshed")
-            #expect(requestCount == 3)
-        }
-    }
-
-    @Test("AuthenticationService propagates refresh task failures")
-    func authenticationServicePropagatesRefreshTaskFailure() async throws {
-        let backend = InMemorySecureStorage()
-        try await withInMemoryBackend(backend) {
-            let host = "refresh-failure-\(UUID().uuidString.lowercased()).example"
-            let namespace = "test.refresh.authentication.failure"
-            let storage = KeychainStorage(namespace: namespace)
-            let session = makeSession(refreshToken: "rt-auth-failure", tokenType: .bearer)
-            let transport = AuthenticationRefreshTransport(responses: [.serverFailure])
-            let storageObserver = RefreshStorageObserver(did: testDID)
-            backend.setOperationObserver { operation, key in
-                storageObserver.observe(operation, key: key)
-            }
-            defer {
-                backend.setOperationObserver(nil)
-                AuthenticationRefreshURLProtocol.uninstall()
-                NetworkService.setNetworkTestProtocolClasses(nil)
-            }
-
-            try await storage.saveAccountAndSession(makeAccount(), session: session, for: testDID)
-            let service = makeAuthenticationService(storage: storage, transport: transport, host: host)
-            let refresh = Task {
-                try await service.refreshTokenIfNeeded(forceRefresh: true)
-            }
-
-            await transport.waitForRequestCount(1)
-            await transport.releaseNextRequest()
-            let result = await refresh.result
-            await storageObserver.refreshCleared.wait()
-
-            switch result {
-            case let .success(unexpectedResult):
-                Issue.record("Refresh failure returned early as \(unexpectedResult) instead of throwing")
-            case let .failure(error):
-                #expect(error as? AuthError == .tokenRefreshFailed)
-            }
-        }
-    }
-
-    @Test("AuthenticationService returns the completed refresh task result")
-    func authenticationServiceReturnsCompletedRefreshTaskResult() async throws {
-        let backend = InMemorySecureStorage()
-        try await withInMemoryBackend(backend) {
-            let host = "refresh-result-\(UUID().uuidString.lowercased()).example"
-            let namespace = "test.refresh.authentication.result"
-            let storage = KeychainStorage(namespace: namespace)
-            let session = makeSession(refreshToken: "rt-auth-result", tokenType: .bearer)
-            let transport = AuthenticationRefreshTransport(responses: [.invalidGrant])
-            let storageObserver = RefreshStorageObserver(did: testDID)
-            backend.setOperationObserver { operation, key in
-                storageObserver.observe(operation, key: key)
-            }
-            defer {
-                backend.setOperationObserver(nil)
-                AuthenticationRefreshURLProtocol.uninstall()
-                NetworkService.setNetworkTestProtocolClasses(nil)
-            }
-
-            try await storage.saveAccountAndSession(makeAccount(), session: session, for: testDID)
-            let service = makeAuthenticationService(storage: storage, transport: transport, host: host)
-            let refresh = Task {
-                try await service.refreshTokenIfNeeded(forceRefresh: true)
-            }
-
-            await transport.waitForRequestCount(1)
-            KeychainManager.clearCache()
-            await transport.releaseNextRequest()
-            let result = try await refresh.value
-            await storageObserver.refreshCleared.wait()
-            await storageObserver.postClearSessionLookup.wait()
-
-            #expect(result == .stillValid, "Refresh did not preserve the task's transient-safe result")
         }
     }
